Bug fixes
Removed that annoying debug println's that would spam whenever you used the smithing anvil, changed up some more things so they use registry names, and made it so when you attack a god you get weakness. It's also supposed to create a beacon beam in the air when they die, but that doesn't work yet so I'll fix that in the future.