Added more blocks to blockstate generator & dungeon command improvements
From using /dungeon a bunch of times I think I've fixed it? I just stopped regenerating the key which seems like a pretty obvious idea to stop issues since it's all in the same frame, I'm not completely sure how this fixed it though so it may be different on a server or something like that.